fix: final non-macro query

+++ crates/banyan-storage-provider-service/src/database/models/block_details.rs
@@ -1,9 +1,8 @@
 #[derive(sqlx::FromRow, Debug)]
 pub struct BlockDetails {
     pub id: String,
-    pub length: i32,
-    pub car_offset: Option<i32>,
+    pub length: i64,
+    pub car_offset: Option<i64>,
     pub base_path: String,
     pub platform_id: String,
-    //-- pub metadata_id: String,
 }
